v.builder: support VTEST_JUST_COMPILE, so `v file_test.v` can just compile a test file to an executable and stop
Normally `v file_test.v` compiles, and then runs, and then removes the compiled executable, after the test was run.
During all that time, the `v` process that compiled it, still stays active and waits for the result.
This can be both wasteful, and also it bloats the time/resource usage measurements.

With VTEST_JUST_COMPILE=1, the v frontend will now *stop and exit* after the compilation, just like it does after:
`v file.v`, *without* executing the produced executable, and without removing it.

That allows a test runner like `v test .`, to measure the test compilation time separately, and if needed, to re-run
the produced executable multiple times (in case of flaky tests), without re-compiling them all the time.
It also allows the test runner, to record the runtime of the test, and its RAM usage separately too,
as well as the size of the produced executable.